requirements.txt
streamlit
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
openai
python-dotenv
groq